Move logappsettingservice
