+++ src/services/EdFi.DataManagementService.Api/Infrastructure/WebApplicationBuilderExtensions.cs
@@ -4,7 +4,6 @@
 // See the LICENSE and NOTICES files in the project root for more information.
 
 using EdFi.DataManagementService.Api.Configuration;
-using Microsoft.Extensions.Options;
 
 namespace EdFi.DataManagementService.Api.Infrastructure
 {
@@ -13,27 +12,8 @@ namespace EdFi.DataManagementService.Api.Infrastructure
         public static void AddServices(this WebApplicationBuilder webAppBuilder)
         {
             webAppBuilder.Services.Configure<AppSettings>(webAppBuilder.Configuration.GetSection("AppSettings"));
-        }
-    }
-
-    // Example class for IOptions 
-    public class LogAppSettingsService()
-    {
-        private readonly ILogger<LogAppSettingsService>? _logger;
-        private readonly IOptions<AppSettings>? _options;
-
-        public LogAppSettingsService(ILogger<LogAppSettingsService>? logger,
-            IOptions<AppSettings>? options) : this()
-        {
-            _logger = logger;
-            _options = options;
-        }
-
-        public void Log()
-        {
-            var appSettings = _options?.Value;
-            _logger?.LogInformation("BeginAllowedSchoolYear = {year}.", appSettings?.BeginAllowedSchoolYear);
-            _logger?.LogInformation("EndAllowedSchoolYear = {year}.", appSettings?.EndAllowedSchoolYear);
+            webAppBuilder.Services.AddLogging(builder => builder.AddConsole());
+            webAppBuilder.Services.AddSingleton<LogAppSettingsService>();
         }
     }
 }
